fix(build): Make the MSVC /W4 /WX build pass
- Define NOMINMAX on MSVC so <windows.h> (via dr_libs) stops clobbering
  std::min/std::max, fixing C2589/C2059 in audio_io.cpp and resample.cpp
- Mark third_party (kissfft/dr_libs/minimp3/r8brain) and the fetched
  Eigen as SYSTEM includes and add /external:W0 so MSVC /W4 /WX no
  longer errors on third-party warnings (C4244/C4127)
- Suppress the intentional alignas-padding warning C4324 via /wd4324
- Add explicit static_cast<int> in cqt.cpp for float-to-int narrowing
  (C4244)

+++ src/feature/cqt.cpp
@@ -373,7 +373,7 @@ Audio icqt(const CqtResult& cqt_result, int length) {
     return Audio();
   }
 
-  int bins_per_octave = constants::kSemitonesPerOctave;
+  int bins_per_octave = static_cast<int>(constants::kSemitonesPerOctave);
   if (frequencies.size() >= 2 && frequencies[0] > 0.0f && frequencies[1] > frequencies[0]) {
     const float ratio = frequencies[1] / frequencies[0];
     const float estimated_bpo = 1.0f / std::log2(ratio);
